Cleaned import and load MgO dataset
Also, I have started experimenting with letting the user write the configuration to a file and seems to be easily doable so I will go for it.
Such testing is what I have written train.py for, still for now is only the starting of such script due to lack of time today.