Cleaned everithing
We are converging to a final form, now I have detached the model used from jax_md simply placing it as a dependencies and not the progect as a fork of the latter.
This allowed for a much cleaner repository that is for much more user friendly and easy to work with also for development.