Added script to perform uncertanty evaluation on md
The mdCommittee script performs uncertanty evaluation over Leopold MD trajectories using the committee approach for acrive learning of neural networks.
In order to handle the huge amount of data created along a FF MD run an approach based on chunking and hdf5 was needed.

Changed Frame description in md.py in order to save the date as Float32 to reduce the amount of memory required for a single trajectory (I'm finishing the sapce on the cluster)