Perfected Training script
Added the evaluation of the average number of neighbours from the training set.
Added a scheduler to the training, using the reduce_on_plateau one.

The md.py script was also changed in order to not account for the absolute value of the magnetization to find the polarons but simply the magnetization per se.